Bug 25861: Removing non-existent 'important' field from SQL
Bug 17232 updated some SQL queries used when adding new MARC
bibliographic frameworks. When backported to 19.11.x, it referenced a
column `important` that does not exist in marc_tag_structure or
marc_subfield_structure in 19.11.x. This patch removes the references to
`important` so that adding new bibliographic frameworks works again.

To test:
1) Ensure you have checked out 19.11.x branch
2) Go to Admin -> MARC bibliographic frameworks
3) Add a new framework and save
4) Go to Actions -> MARC structure for your new framework
5) Use the 'create framework using' dropdown to import the structure
from an existing framework
6) Note that the structure does not populate, and there is now an error
in the logs "Unknown column 'important'"
7) Apply this patch and restart services
8) Repeat steps 3-5
9) Confirm the structure is populated correctly and there is no error in
the logs

Sponsored-by: Catalyst IT

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>

Signed-off-by: Aleisha Amohia <aleishaamohia@hotmail.com>