Bug 25798: Add copyright year to "Holds to pull" report
Display copyright year (publication year) of book
together with title in "Holds to pull" report.

Copyright year is fetched from biblio -table as "copyrightyear"
and displayed when the preferred marc is set to MARC21.
Publication year is fetched from biblioitems -table as "publicationyear"
and displayed when the preferred marc is set to UNIMARC.

1. Place a hold on book with known copyright year.
2. Go to /cgi-bin/koha/circ/pendingreserves.pl and check the "title"
table of that book that you placed hold on.
3. Observe that there's no information about publication year in that
field.
4. Apply patch.
5. Repeat step 2.
6. Observe that copyright year (publication year) of book appeared
in the title table after book's title and author.

Mentored-by: Andrew Nugged <nugged@gmail.com>

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>