Bug 25827: Add floating toolbar to the holds summary page in staff interface
This patch adds a floating toolbar to the holds summary page in the staff
interface.

To test, apply the patch and view the holds page for a title with
multiple holds on it. Test that when you scroll down past the position
of the "Update hold(s)" button, the toolbar "sticks" to the top of the
screen.

Test that holds are updated correctly after clicking the "Update
hold(s)" button.

Signed-off-by: Emmi Takkinen <emmi.takkinen@outlook.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>