Bug 22773: (QA follow-up) Improve existing merge button and fix checkbox label
Fixes the checkbox label by adding an id, so that when you click on
"Show only subscriptons" the checkbox will be checked.

Adds standard classes to the existing "Merge selected invoices" button
to make things more consistent. Also added the fa-compress icon that
we use in cataloguing on the merge records button.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>