Bug 25987: Fix alignment of fields in label batch creation form
This patch makes minor markup changes to the template for the label
batch creation process so that form fields line up correctly.

Unrelated change: A hint has been moved out of the label area to below
the textarea in order to conform with established patterns.

To test, apply the patch and go to Tools > Label creator > New > Label
batch.

 - In the batch creation form the "Enter by itemnumber" and "Enter by
   barcode" should be aligned correctly with each other.
 - The "One number per line..." hint should appear below the textarea.
 - Test that the form fields are still correctly aligned after adding
   some items to the batch and the "description" field is shown in the
   form.

Works for me!

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>