Bug 24986: DBIC Schema files + fix of bug 23081 backport
Bug 23081 backported the DBIC changes instead of regenerating the files.
Thus the checksum weren't matching because 19.05 also had other changes
to the schema so the same changes result in a different checksum.
The small changes were manually undone and the DBIC were regenerated.
So this commit contains the changes for both patchsets.

Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>